Remove unnecessary methods and variables
