+++ src/DominationRound.js
@@ -1,14 +1,10 @@
 import Round from './Round.js';
-import DominationTurn from './DominationTurn.js';
 import DOMupdates from "./DOMupdates";
 
 class DominationRound extends Round {
   constructor(game, multiplier) {
     super(game);
     this.multiplier = multiplier;
-    this.seconds = 30;
-    this.respondentsInfo = [];
-    this.dominationGuesses = [];
   }
 
   endDominationRound() {
