+++ src/RegularTurn.js
@@ -1,4 +1,3 @@
-import Round from "./Round";
 import DOMupdates from "./DOMupdates";
 import Turn from "./Turn"
 
@@ -9,22 +8,10 @@ class RegularTurn extends Turn {
     this.currentPlayer = currentRound.currentPlayer;
   }
 
-  checkGuess(guess) {
-    return this.currentRound.answerStrings.includes(guess.toUpperCase());
-  }
-
   spliceAnswers(guess) {
     let i = this.currentRound.answerStrings.findIndex(answer => answer === guess.toUpperCase())
     this.currentRound.answerStrings[i] = 'false';
     this.currentRound.endRound();
-  }
-
-  countRespondents(guess) {
-    let answerInfo = this.currentRound.answers.find(answer => {
-      let textAnswer = answer.answer.toUpperCase();
-      return textAnswer.includes(guess.toUpperCase())
-    })
-    return answerInfo.respondents;
   }
 
   updateScore(guess) {
