+++ src/Turn.js
@@ -1,4 +1,3 @@
-import Round from "./Round";
 import DOMupdates from "./DOMupdates";
 
 class Turn {
@@ -8,11 +7,19 @@ class Turn {
   }
 
   checkGuess(guess) {
-    return this.currentRound.answerStrings.includes(this.capitalizeGuess(guess));
+    return this.currentRound.answerStrings.includes(guess.toUpperCase());
   }
 
   giveFeedback(guess) {
     DOMupdates.displayFeedback(this.checkGuess(guess));
+  }
+
+  countRespondents(guess) {
+    let answerInfo = this.currentRound.answers.find(answer => {
+      let textAnswer = answer.answer.toUpperCase();
+      return textAnswer.includes(guess.toUpperCase())
+    })
+    return answerInfo.respondents;
   }
 }
